+++ forms/src/misc/tokens.ts
@@ -1,0 +1,12 @@
+import {InjectionToken} from '@angular/core';
+import {StringDictionary} from '@jscrpt/common';
+
+/**
+ * Injection token containing validation error messages
+ */
+export const VALIDATION_ERROR_MESSAGES: InjectionToken<StringDictionary> = new InjectionToken<StringDictionary>('VALIDATION_ERROR_MESSAGES');
+
+/**
+ * Injection token for injecting array of ignored error names when automatically processed
+ */
+export const IGNORED_VALIDATION_ERRORS: InjectionToken<string[]> = new InjectionToken<string[]>('IGNORED_VALIDATION_ERRORS', {providedIn: 'root', factory: () => ['actual']});